cleanup ; add train/test scripts; comment code
