--- grobid/58102005.results
@@ -1,49 +1,0 @@
-===== Token-level results =====
-
-
-label           accuracy        precision       recall          f1
-
-<author>                98.47           99.77           94.9            97.28
-<title>         99.67           99.86           99.04           99.45
-<editor>                99.86           98.05           83.97           90.46
-<booktitle>             99.91           98.91           99.06           98.98
-<date>          99.99           99.88           99.8            99.84
-<journal>               99.75           99.37           96.3            97.81
-<volume>                99.93           98.48           97.84           98.16
-<issue>         99.96           83.13           94.39           88.4
-<tech>          100             99.48           100             99.74
-<institution>   99.98           95.71           99.79           97.71
-<pages>         99.84           99.84           97.84           98.83
-<location>              99.9            91              98.36           94.54
-<publisher>             99.97           95.85           97.88           96.86
-<note>          99.95           81.51           92.61           86.7
-<web>           99.99           100             91.13           95.36
-<pubnum>                99.99           79.41           100             88.52
-
-all fields (micro)              99.82           99.46           97.15           98.29
-all fields (macro)              99.82           95.02           96.43           95.54
-
-===== Field-level results =====
-
-label           accuracy        precision       recall          f1
-
-<author>                94.86           98.64           42.89           59.78
-<title>         99.56           99.06           95.89           97.45
-<editor>                99.89           97.18           63.3            76.67
-<booktitle>             99.91           96.86           97.62           97.24
-<date>          99.97           99.86           99.86           99.86
-<journal>               99.39           99.62           91.44           95.35
-<volume>                99.83           98.47           99.11           98.79
-<issue>         99.87           82.85           95.19           88.59
-<tech>          99.99           97.44           100             98.7
-<institution>   99.98           92.86           98.73           95.71
-<pages>         99.98           99.73           99.97           99.85
-<location>              99.77           84.73           98.48           91.09
-<publisher>             99.94           94.88           97.89           96.36
-<note>          99.91           69.05           86.57           76.82
-<web>           100             100             100             100
-<pubnum>                100             80              100             88.89
-
-all fields (micro)              99.55           98.31           88.11           92.93
-all fields (macro)              99.55           93.2            91.68           91.32
-
