--- grobid/58102005.results.field
@@ -1,22 +1,0 @@
-
-label           accuracy        precision       recall          f1
-
-<author>                94.86           98.64           42.89           59.78
-<title>         99.56           99.06           95.89           97.45
-<editor>                99.89           97.18           63.3            76.67
-<booktitle>             99.91           96.86           97.62           97.24
-<date>          99.97           99.86           99.86           99.86
-<journal>               99.39           99.62           91.44           95.35
-<volume>                99.83           98.47           99.11           98.79
-<issue>         99.87           82.85           95.19           88.59
-<tech>          99.99           97.44           100             98.7
-<institution>   99.98           92.86           98.73           95.71
-<pages>         99.98           99.73           99.97           99.85
-<location>              99.77           84.73           98.48           91.09
-<publisher>             99.94           94.88           97.89           96.36
-<note>          99.91           69.05           86.57           76.82
-<web>           100             100             100             100
-<pubnum>                100             80              100             88.89
-
-all fields (micro)              99.55           98.31           88.11           92.93
-all fields (macro)              99.55           93.2            91.68           91.32
