--- grobid/58102005.results.token
@@ -1,23 +1,0 @@
-
-
-label           accuracy        precision       recall          f1
-
-<author>                98.47           99.77           94.9            97.28
-<title>         99.67           99.86           99.04           99.45
-<editor>                99.86           98.05           83.97           90.46
-<booktitle>             99.91           98.91           99.06           98.98
-<date>          99.99           99.88           99.8            99.84
-<journal>               99.75           99.37           96.3            97.81
-<volume>                99.93           98.48           97.84           98.16
-<issue>         99.96           83.13           94.39           88.4
-<tech>          100             99.48           100             99.74
-<institution>   99.98           95.71           99.79           97.71
-<pages>         99.84           99.84           97.84           98.83
-<location>              99.9            91              98.36           94.54
-<publisher>             99.97           95.85           97.88           96.86
-<note>          99.95           81.51           92.61           86.7
-<web>           99.99           100             91.13           95.36
-<pubnum>                99.99           79.41           100             88.52
-
-all fields (micro)              99.82           99.46           97.15           98.29
-all fields (macro)              99.82           95.02           96.43           95.54
