--- grobid/fieldLevelGrobid.eval
@@ -1,22 +1,0 @@
-
-label		accuracy	precision	recall		f1
-
-<author>		88.78		36.27		69.43		47.65
-<title>		95.48		64.29		82.78		72.37
-<editor>		99.76		0		0		0
-<pages>		94.36		53.58		96.65		68.94
-<booktitle>		98.61		44.19		6.02		10.6
-<date>		96.74		99.07		57.54		72.8
-<journal>		97.72		82.79		75		78.7
-<volume>		96.15		100		31.52		47.93
-<issue>		99.55		0		0		0
-<tech>		99.84		0		0		0
-<institution>	99.83		0		0		0
-<location>		99		0		0		0
-<publisher>		99.38		0		0		0
-<note>		99.85		0		0		0
-<web>		99.98		0		0		0
-<pubnum>		99.99		0		0		0
-
-all fields (micro)		97.81		59.71		63.01 		61.32
-all fields (macro)		97.81		30.01		26.18		24.94
