--- grobid/tokenLevelGrobid.eval
@@ -1,22 +1,0 @@
-
-label		accuracy	precision	recall		f1
-
-<author>		89.67		70.81		97.07		81.89
-<title>		91.92		80.64		89.28		84.74
-<editor>		99.33		0		0		0
-<booktitle>		96.38		51.61		1.1		2.15
-<pages>		97.56		73.34		89.56		80.65
-<date>		98.73		99.07		49.83		66.3
-<journal>		97.77		90.43		60.4		72.43
-<volume>		98.87		100		30.96		47.28
-<issue>		99.87		0		0		0
-<tech>		99.76		0		0		0
-<institution>	99.71		0		0		0
-<location>		99.28		0		0		0
-<publisher>		99.59		0		0		0
-<note>		99.84		0		0		0
-<web>		99.92		0		0		0
-<pubnum>		99.98		0		0		0
-
-all fields (micro)		98.01		76.35		79.2 		77.75
-all fields (macro)		98.01		35.37		26.14		27.21
