--- myresults/36112005.results
@@ -1,51 +1,0 @@
-===== Token-level results =====
-
-
-label           accuracy        precision       recall          f1
-
-author          98.23           99.76           93.99           96.79
-other           97.22           84.78           96.57           90.29
-title           99.56           99.76           98.76           99.26
-editor          99.84           95.48           84.15           89.46
-booktitle               99.79           97.40           97.62           97.51
-date            99.88           99.24           96.74           97.97
-journal         98.78           85.36           94.90           89.88
-volume          99.72           89.55           96.71           92.99
-issue           99.84           37.50           4.21            7.56
-tech            99.93           98.66           77.02           86.51
-institution             99.88           94.75           69.15           79.95
-pages           99.68           99.48           95.66           97.53
-location                99.89           90.65           96.82           93.63
-publisher               99.96           95.19           95.76           95.48
-note            99.88           67.94           69.26           68.59
-web             99.92           100.00          16.13           27.78
-pubnum          99.99           79.41           100.00          88.52
-
-all fields (micro)              99.53           95.84           95.84           95.84
-all fields (macro)              99.53           89.11           81.38           82.34
-
-
-===== Field-level results =====
-
-label           accuracy        precision       recall          f1
-
-author          94.06           97.13           32.87           49.12
-other           92.38           85.13           97.62           90.95
-title           99.50           98.65           95.37           96.98
-editor          99.89           93.51           66.06           77.42
-booktitle               99.82           94.17           94.77           94.47
-date            99.75           99.36           97.89           98.62
-journal         99.08           95.04           90.90           92.92
-volume          99.10           89.51           97.96           93.54
-issue           99.45           37.50           4.33            7.76
-tech            99.92           92.45           64.47           75.97
-institution             99.92           91.38           67.09           77.37
-pages           99.72           99.35           97.02           98.17
-location                99.74           83.58           96.97           89.78
-publisher               99.93           93.88           97.18           95.50
-note            99.85           55.42           68.66           61.33
-web             99.98           100.00          18.18           30.77
-pubnum          100.00          80.00           100.00          88.89
-
-all fields (micro)              98.95           90.51           89.82           90.16
-all fields (macro)              98.95           87.42           75.72           77.62
