--- myresults/36112005.results.field
@@ -1,22 +1,0 @@
-
-label           accuracy        precision       recall          f1
-
-author          94.06           97.13           32.87           49.12
-title           99.50           98.65           95.37           96.98
-editor          99.89           93.51           66.06           77.42
-booktitle               99.82           94.17           94.77           94.47
-date            99.75           99.36           97.89           98.62
-journal         99.08           95.04           90.90           92.92
-volume          99.10           89.51           97.96           93.54
-issue           99.45           37.50           4.33            7.76
-tech            99.92           92.45           64.47           75.97
-institution             99.92           91.38           67.09           77.37
-pages           99.72           99.35           97.02           98.17
-location                99.74           83.58           96.97           89.78
-publisher               99.93           93.88           97.18           95.50
-note            99.85           55.42           68.66           61.33
-web             99.98           100.00          18.18           30.77
-pubnum          100.00          80.00           100.00          88.89
-
-all fields (micro)              98.95           90.51           89.82           90.16
-all fields (macro)              98.95           87.42           75.72           77.62
