--- myresults/36112005.results.token
@@ -1,22 +1,0 @@
-
-label           accuracy        precision       recall          f1
-
-author          98.23           99.76           93.99           96.79
-title           99.56           99.76           98.76           99.26
-editor          99.84           95.48           84.15           89.46
-booktitle               99.79           97.40           97.62           97.51
-date            99.88           99.24           96.74           97.97
-journal         98.78           85.36           94.90           89.88
-volume          99.72           89.55           96.71           92.99
-issue           99.84           37.50           4.21            7.56
-tech            99.93           98.66           77.02           86.51
-institution             99.88           94.75           69.15           79.95
-pages           99.68           99.48           95.66           97.53
-location                99.89           90.65           96.82           93.63
-publisher               99.96           95.19           95.76           95.48
-note            99.88           67.94           69.26           68.59
-web             99.92           100.00          16.13           27.78
-pubnum          99.99           79.41           100.00          88.52
-
-all fields (micro)              99.53           95.84           95.84           95.84
-all fields (macro)              99.53           89.11           81.38           82.34
