--- myresults/55092005.results
@@ -1,50 +1,0 @@
-===== Token-level results =====
-
-
-label           accuracy        precision       recall          f1
-
-author          98.24           99.79           93.98           96.80
-other           97.24           84.83           96.61           90.34
-title           99.57           99.77           98.76           99.26
-editor          99.85           95.50           84.62           89.73
-booktitle               99.79           97.49           97.64           97.56
-date            99.88           99.24           96.74           97.97
-journal         98.78           85.34           95.01           89.92
-volume          99.72           89.59           96.78           93.05
-issue           99.84           37.50           4.21            7.56
-tech            99.93           98.67           77.28           86.68
-institution             99.88           94.75           69.15           79.95
-pages           99.68           99.49           95.66           97.54
-location                99.90           91.40           96.90           94.07
-publisher               99.96           95.21           96.22           95.71
-note            99.88           66.18           70.82           68.42
-web             99.92           100.00          16.13           27.78
-pubnum          99.99           79.41           100.00          88.52
-
-all fields (micro)              99.53           95.86           95.86           95.86
-all fields (macro)              99.53           89.07           81.56           82.40
-
-===== Field-level results =====
-
-label           accuracy        precision       recall          f1
-
-author          94.07           97.38           32.87           49.15
-other           92.40           85.17           97.64           90.98
-title           99.49           98.59           95.40           96.97
-editor          99.89           93.51           66.06           77.42
-booktitle               99.83           94.47           94.77           94.62
-date            99.75           99.36           97.89           98.62
-journal         99.08           95.04           90.90           92.92
-volume          99.10           89.54           98.00           93.58
-issue           99.45           37.50           4.33            7.76
-tech            99.92           92.59           65.79           76.92
-institution             99.92           91.38           67.09           77.37
-pages           99.72           99.38           97.02           98.19
-location                99.75           84.24           97.19           90.25
-publisher               99.93           93.88           97.18           95.50
-note            99.85           54.55           71.64           61.94
-web             99.98           100.00          18.18           30.77
-pubnum          100.00          80.00           100.00          88.89
-
-all fields (micro)              98.95           90.54           89.84           90.19
-all fields (macro)              98.95           87.45           76.00           77.76
