--- myresults/55092005.results.field
@@ -1,22 +1,0 @@
-
-label           accuracy        precision       recall          f1
-
-author          94.07           97.38           32.87           49.15
-title           99.49           98.59           95.40           96.97
-editor          99.89           93.51           66.06           77.42
-booktitle               99.83           94.47           94.77           94.62
-date            99.75           99.36           97.89           98.62
-journal         99.08           95.04           90.90           92.92
-volume          99.10           89.54           98.00           93.58
-issue           99.45           37.50           4.33            7.76
-tech            99.92           92.59           65.79           76.92
-institution             99.92           91.38           67.09           77.37
-pages           99.72           99.38           97.02           98.19
-location                99.75           84.24           97.19           90.25
-publisher               99.93           93.88           97.18           95.50
-note            99.85           54.55           71.64           61.94
-web             99.98           100.00          18.18           30.77
-pubnum          100.00          80.00           100.00          88.89
-
-all fields (micro)              98.95           90.54           89.84           90.19
-all fields (macro)              98.95           87.45           76.00           77.76
