--- myresults/55092005.results.token
@@ -1,21 +1,0 @@
-label           accuracy        precision       recall          f1
-
-author          98.24           99.79           93.98           96.80
-title           99.57           99.77           98.76           99.26
-editor          99.85           95.50           84.62           89.73
-booktitle               99.79           97.49           97.64           97.56
-date            99.88           99.24           96.74           97.97
-journal         98.78           85.34           95.01           89.92
-volume          99.72           89.59           96.78           93.05
-issue           99.84           37.50           4.21            7.56
-tech            99.93           98.67           77.28           86.68
-institution             99.88           94.75           69.15           79.95
-pages           99.68           99.49           95.66           97.54
-location                99.90           91.40           96.90           94.07
-publisher               99.96           95.21           96.22           95.71
-note            99.88           66.18           70.82           68.42
-web             99.92           100.00          16.13           27.78
-pubnum          99.99           79.41           100.00          88.52
-
-all fields (micro)              99.53           95.86           95.86           95.86
-all fields (macro)              99.53           89.07           81.56           82.40
