--- myresults/tokenLevelMine.eval
@@ -1,16 +1,0 @@
-
-label		accuracy	precision	recall		f1
-
-B-author		100.00		100.00		100.00		100.00
-author		100.00		100.00		100.00		100.00
-title		100.00		100.00		100.00		100.00
-B-other		100.00		100.00		100.00		100.00
-pages		100.00		100.00		100.00		100.00
-booktitle		100.00		100.00		100.00		100.00
-journal		100.00		100.00		100.00		100.00
-B-pages		100.00		100.00		100.00		100.00
-B-date		100.00		100.00		100.00		100.00
-other		100.00		100.00		100.00		100.00
-
-all fields (micro)		100.00		100.00		100.00 		100.00
-all fields (macro)		100.00		100.00		100.00		100.00
