+++ src/main/scala/edu/umass/cs/iesl/bibie/Grobid.scala
@@ -4,12 +4,197 @@ package edu.umass.cs.iesl.bibie
  * Created by kate on 5/13/15.
  */
 
+import java.io._
+import java.net.URL
+
 import cc.factorie.app.nlp._
-import cc.factorie.app.nlp.segment.DeterministicTokenizer
+import cc.factorie.optimize._
+import cc.factorie.util._
 import scala.io.Source
-import java.io.File
-import scala.xml.{NodeSeq, Node}
-import scala.collection.mutable.{ArrayBuffer, HashMap}
+import scala.collection.mutable.ArrayBuffer
+
+/**
+ * Infrastructure for comparison of GROBID vs IESL.
+ * We load documents from a file generated by GROBID's training/eval process, which contains (a) all the features and
+ * feature values GROBID uses and (b) gold labels. This way, we can more or less directly compare our learning vs.
+ * GROBID's learning. One thing to note however is that GROBID uses a cascade of CRF's (I'm not 100% sure how this works
+ * yet). This might be a TODO to consider implementing.
+ */
+
+/**
+ * Train a model using GROBID data and features.
+ */
+class GrobidCitationCRFTrainer {
+  val evaluator = new SegmentationEvaluation[CitationLabel](LabelDomain)
+  val evaluator2 = new ExactlyLikeGrobidEvaluator
+  val model = new CitationCRFModel
+  def process(document: Document): Document = {
+    if (document.tokens.size == 0) return document
+    for (sentence <- document.sentences if sentence.tokens.size > 0) {
+      val vars = sentence.tokens.map(_.attr[CitationLabel]).toSeq
+      val sum = CitationBIOHelper.infer(vars, model)
+      sum.setToMaximize(null)
+    }
+    document
+  }
+  def train(trainDocuments: Seq[Document], testDocuments: Seq[Document], params: Hyperparams): Unit = {
+    implicit val random = new scala.util.Random
+    // make "tmp" directory in root project folder ; needed for evaluations
+    val tmpDir = params.root + "/tmp"
+    if (!(new java.io.File(tmpDir).exists)) {
+      import scala.sys.process._
+      assert(s"mkdir $tmpDir".! == 0, s"could not mkdir $tmpDir")
+    }
+    // Get the variables to be inferred (for now, just operate on a subset)
+    val trainLabels: Seq[CitationLabel] = trainDocuments.flatMap(_.tokens).map(_.attr[CitationLabel]).toSeq
+    val testLabels: Seq[CitationLabel] = testDocuments.flatMap(_.tokens).map(_.attr[CitationLabel]).toSeq
+    (trainLabels ++ testLabels).filter(_ != null).foreach(_.setRandomly(random))
+    val vars = for (td <- trainDocuments; sentence <- td.sentences if sentence.length > 1) yield sentence.tokens.map(_.attr[CitationLabel])
+    val examples = vars.map(v => new LikelihoodExample(v.toSeq, model, cc.factorie.infer.InferByBPChain))
+    def evaluate(): Unit = {
+      val tmpFile = tmpDir + "/" + scala.util.Random.nextInt(10000) + ".tmp"
+      trainDocuments.foreach(process)
+      val (trainF1, _) = evaluator2.evaluate(trainDocuments, tmpFile)
+      println(s"TRAIN F1 = $trainF1")
+      testDocuments.foreach(process)
+      val (testF1, _) = evaluator2.evaluate(testDocuments, tmpFile)
+      println(s"DEV F1 = $testF1")
+    }
+    params.optimizer match {
+      case "lbfgs" =>
+        val optimizer = new LBFGS(maxIterations=1814) with L2Regularization
+        val trainer = new ThreadLocalBatchTrainer(model.parameters, optimizer)
+        println("training with LBFGS ...")
+        trainer.trainFromExamples(examples)
+      case "adagrad" =>
+        val optimizer = new AdaGradRDA(delta=params.delta, rate=params.rate, l1=params.l1, l2=params.l2, numExamples=examples.length)
+        println("training with AdaGradRDA ...")
+        Trainer.onlineTrain(model.parameters, examples, evaluate=evaluate, useParallelTrainer=false, maxIterations=5, optimizer=optimizer)
+      case _ => throw new Exception(s"invalid optimizer: ${params.optimizer}")
+    }
+    (trainLabels ++ testLabels).foreach(_.setRandomly(random))
+    trainDocuments.foreach(process)
+    testDocuments.foreach(process)
+    evaluate()
+    evaluator.printEvaluation(testDocuments, testDocuments, "FINAL")
+  }
+
+  def serialize(stream: OutputStream) {
+    import cc.factorie.util.CubbieConversions._
+    val is = new DataOutputStream(stream)
+    BinarySerializer.serialize(LabelDomain, is)
+    BinarySerializer.serialize(CitationFeaturesDomain.dimensionDomain, is)
+    BinarySerializer.serialize(model, is)
+    is.close()
+  }
+  def deserialize(stream: InputStream): Unit = deserialize(model, stream)
+
+  def deserialize(theModel: CitationCRFModel, stream: InputStream) {
+    import cc.factorie.util.CubbieConversions._
+    val is = new DataInputStream(stream)
+    BinarySerializer.deserialize(LabelDomain, is)
+    LabelDomain.freeze()
+    println("deserialized labeldomain")
+    BinarySerializer.deserialize(CitationFeaturesDomain.dimensionDomain, is)
+    CitationFeaturesDomain.freeze()
+    println("deserialized featuresdomain")
+    BinarySerializer.deserialize(theModel, is)
+    println("deserialized model")
+    is.close()
+  }
+}
+
+object TrainCitationModelGrobid extends HyperparameterMain {
+  URLHandlerSetup.poke()
+  def initFeatures(docs: Seq[Document]): Unit = {
+    docs.flatMap(_.tokens).foreach { token =>
+      token.attr += new CitationFeatures(token)
+      token.attr[CitationFeatures] ++= token.attr[PreFeatures].features
+    }
+  }
+  def evaluateParameters(args: Array[String]): Double = {
+    println(args.mkString(", "))
+    val opts = new TrainCitationModelOptions
+    opts.parse(args)
+    val params = new Hyperparams(opts)
+    val trainer = new GrobidCitationCRFTrainer
+    val allData = LoadGrobid.fromFilename(opts.trainFile.value)
+    val trainPortion = (allData.length.toDouble * opts.trainPortion.value).floor.toInt
+    val trainingData = allData.take(trainPortion)
+    initFeatures(trainingData)
+    CitationFeaturesDomain.freeze()
+    val devData = allData.drop(trainPortion)
+    initFeatures(devData)
+    println(s"feature domain size: ${CitationFeaturesDomain.dimensionSize}")
+    println(s"training data: ${trainingData.length} docs, ${trainingData.flatMap(_.tokens).length} tokens")
+    println(s"dev data: ${devData.length} docs, ${devData.flatMap(_.tokens).length} tokens")
+    trainer.train(trainingData, devData, params)
+    val evaluator = new ExactlyLikeGrobidEvaluator(opts.rootDir.value)
+    val (f0, eval) = evaluator.evaluate(devData, opts.outputFile.value, writeFiles=opts.writeEvals.value, outputDir=opts.outputDir.value)
+    println(eval)
+    if (opts.saveModel.value) {
+      println(s"serializing model to: ${opts.modelFile.value}")
+      trainer.serialize(new FileOutputStream(opts.modelFile.value))
+    }
+    f0
+  }
+}
+
+object OptimizeCitationModelGrobid {
+  def main(args: Array[String]): Unit = {
+    println(args.mkString(", "))
+    val opts = new TrainCitationModelOptions
+    opts.parse(args)
+    opts.saveModel.setValue(false)
+    opts.writeEvals.setValue(false)
+    val l1 = HyperParameter(opts.l1, new LogUniformDoubleSampler(1e-6, 1))
+    val l2 = HyperParameter(opts.l2, new LogUniformDoubleSampler(1e-6, 1))
+    val qs = new QSubExecutor(10, "edu.umass.cs.iesl.bibie.TrainCitationModelGrobid")
+    val optimizer = new HyperParameterSearcher(opts, Seq(l1, l2), qs.execute, 200, 180, 60)
+    val result = optimizer.optimize()
+    println("Got results: " + result.mkString(" "))
+    println("Best l1: " + opts.l1.value + " best l2: " + opts.l2.value)
+    println("Running best configuration...")
+    opts.saveModel.setValue(true)
+    opts.writeEvals.setValue(true)
+    import scala.concurrent.duration._
+    import scala.concurrent.Await
+    Await.result(qs.execute(opts.values.flatMap(_.unParse).toArray), 1.hours)
+    println("Done.")
+  }
+}
+
+object TestCitationModelGrobid {
+  URLHandlerSetup.poke()
+  def initFeatures(docs: Seq[Document]): Unit = {
+    docs.flatMap(_.tokens).foreach { token =>
+      token.attr += new CitationFeatures(token)
+      token.attr[CitationFeatures] ++= token.attr[PreFeatures].features
+    }
+  }
+  def main(args: Array[String]): Unit = {
+    val opts = new TrainCitationModelOptions
+    opts.parse(args)
+    println(s"deserializing model from: ${opts.modelFile.value}")
+    val trainer = new GrobidCitationCRFTrainer
+    trainer.deserialize(new URL(opts.modelFile.value).openStream())
+    CitationFeaturesDomain.freeze()
+    println(s"loading file: ${opts.testFile.value}")
+    val data = LoadGrobid.fromFilename(opts.testFile.value)
+    initFeatures(data)
+    println("processing docs")
+    data.foreach(trainer.process)
+    println("evaluating")
+    val evaluator = new ExactlyLikeGrobidEvaluator(opts.rootDir.value)
+    println(evaluator.evaluate(data, opts.outputFile.value, writeFiles=opts.writeEvals.value, outputDir=opts.outputDir.value))
+    println("done.")
+  }
+}
+
+
+
+class GoldCitationLabel(val label: String, val token: Token)
+class PreFeatures(val features: Array[String], val token: Token)
 
 /*
 M m M M M M M M M M LINESTART ALLCAP NODIGIT 1 0 0 0 0 0 0 0 0 0 0 0 NOPUNCT 1 I-<author>
@@ -27,9 +212,6 @@ Moto moto M Mo Mot Moto o to oto Moto LINEIN INITCAP NODIGIT 0 0 0 0 0 0 0 0 0 0
 - - - - - - - - - - LINEIN ALLCAP NODIGIT 1 0 0 0 0 0 0 0 0 0 0 0 HYPHEN 4 <author>
 oka oka o ok oka oka a ka oka oka LINEIN NOCAPS NODIGIT 0 1 0 0 0 0 0 0 0 0 0 0 NOPUNCT 5 <author>
  */
-
-class GoldCitationLabel(val label: String, val token: Token)
-class PreFeatures(val features: Array[String], val token: Token)
 
 object LoadGrobid {
   def fromFilename(filename: String): Seq[Document] = {
@@ -136,60 +318,10 @@ object LoadGrobid {
         }
       }
     }
-
     println(s"Loaded $docCount docs with $tokenCount tokens from file $filename.")
     buff
   }
-
-
 }
 
-/*
-<ref-marker> [30] </ref-marker> <authors> <person> <person-first> E. </person-first> <person-middle> W. </person-middle> <person-last> Montroll , </person-last> </person> <person> <person-first> B. </person-first> <person-middle> J. </person-middle> <person-last> West , </person-last> </person> </authors> <venue> <booktitle> Fluctuation Phenomena , </booktitle> <publisher> Elsevier Science Publishers B. V. , </publisher> <address> Amsterdam , </address> <date> <year> 1979 , </year> </date> <chapter> Ch . On an enriched collection of stochastic processes , </chapter> <pages> pp . 61--205 . </pages> </venue>
- */
-
-/*
-  def fromFile(filename: String): Seq[Document] = {
-    var document: Document = null
-    val documents = new collection.mutable.ArrayBuffer[Document]
-    var sentence: Sentence = null
-
-    for (l <- Source.fromFile(filename).getLines()) {
-      document = new Document("").setName("NLP-" + documents.length)
-      documents += document
-      sentence = new Sentence(document)
-      val tags = new mutable.HashMap[String, (Int, Int)]()
-      var count = 0
-      for (t <- l.split("\\s+")) {
-        count += 1
-        if (t.matches("<[a-z\\-_]*>")) {tags(t.replaceAll("<|>", "")) = (count, 0)}
-        else if (t.matches("</[a-z\\-_]*>")) tags.remove(t.replaceAll("<|>|/", ""))
-        else {
-
-          // TODO fixme base model is trained on "" not "O"
-          val tag = /*if (tags.isEmpty) "O" else */ tags.toSeq.sortBy(_._2._1).map(ta => (if (ta._2._2 == 0) "B-" else "I-") + ta._1).mkString(":")
-          if (Exclude.ar.contains(tag)) {
-            println("problem with" + tag)
-            println("on token" + t)
-            println("Here: " + sentence.tokens.map(_.string).mkString(" "))
-            println("===========")
-          }
-          /*if(tag == "I-venue") {
-            println("Empty tag: " + tag)
-            println("on token" + t)
-            println("Here: " + sentence.tokens.map(_.string).mkString(" "))
-            println("===========")
-          }*/
-          if (sentence.length > 0) document.appendString(" ")
-          val token = new Token(sentence, t)
-          token.attr += new CitationLabel(if (!LabelDomain.frozen || LabelDomain.categories.contains(tag)) tag else "O", token)
-          tags.foreach {t => tags(t._1) = (t._2._1, t._2._2 + 1)}
-        }
-      }
-    }
-    println("Loaded: " + documents.size + " number of citations.")
-    documents.toSeq
-  }
- */
